replace DecimalFields with FloatFields in forms

+++ analysis/irises_analysis/form_add.py
@@ -6,10 +6,10 @@ class Form_add(forms.Form):
             (2, 'Versicolour'),
             (3, 'Virginica'),
         ]
-    sepal_length = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal length'}), label='')
-    sepal_width = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal width'}), label='')
-    petal_length = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal length'}), label='')
-    petal_width = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal width'}), label='')
+    sepal_length = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal length'}), label='')
+    sepal_width = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal width'}), label='')
+    petal_length = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal length'}), label='')
+    petal_width = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal width'}), label='')
     iris_class = forms.ChoiceField(
         label="Class",
         choices=CLASS_CHOICES,
@@ -17,7 +17,7 @@ class Form_add(forms.Form):
     )
 
 class Form_predict(forms.Form):
-    sepal_length = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal length'}), label='')
-    sepal_width = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal width'}), label='')
-    petal_length = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal length'}), label='')
-    petal_width = forms.DecimalField(decimal_places=2, widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal width'}), label='')
+    sepal_length = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal length'}), label='')
+    sepal_width = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'sepal width'}), label='')
+    petal_length = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal length'}), label='')
+    petal_width = forms.FloatField(widget=forms.TextInput(attrs={'class': 'form-control', 'placeholder': 'petal width'}), label='')
